Run acceptance test primarily aginst in-memory engine
As a general rule, I think that when testing the logic of a dataset
definition we should do so against the in-memory engine. Ensuring that
this accurately reflects the behaviour of the SQL engines is a job for a
different set of tests (the spec and generative tests).

However I've retained the full TPP backend test for now as the current
logic test is so anemic.